+++ Makefile
@@ -14,7 +14,7 @@ SRC_DIR = src
 TEST_DIR = test/build
 REPORTS_DIR = $(TEST_DIR)/reports
 
-SRC_FILES = bufcache.c buffer.c cdllist.c diskdrv.c util.c talloc.c inode.c inocache.c bfs_drv.c su_blk.c
+SRC_FILES = bufcache.c buffer.c cdllist.c diskdrv.c util.c talloc.c inode.c inocache.c su_blk.c bfs.c driver.c
 OBJS_SRC = $(patsubst %.c, %.o, $(SRC_FILES))
 OBJS = main.o $(OBJS_SRC)
 BUILD_OBJS = $(addprefix $(BUILD_DIR)/,$(OBJS))
