+++ test/test.vfs
@@ -10,5 +10,5 @@ test disk block 3 content here
 test disk block 4 content here
 test disk block 5 content here
 test disk block 6 content here
-test disk block 7 content here
+Hello worldlock 7 content here
 test disk block 8 content here
